+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * GUI
 
   * Updated ImGui to 1.69: [#1274](https://github.com/dartsim/dart/pull/1274)
+  * Add VoxelGridShape support to OSG renderer: [#1276](https://github.com/dartsim/dart/pull/1276)
 
 * Build System